+++ packages/ub/modules/Session.js
@@ -113,8 +113,8 @@ Object.defineProperty(Session, 'userLang', {
  * @property {string} roles Logged in user roles names separated by comma. In most case better to use uData.roleIDs array. Added by `ub` model
  * @property {Array<number>} roleIDs Array or role IDs for logged in user roles. Added by `ub` model
  * @property {Array<number>} groupIDs Array or group IDs for logged in user roles. Added by `ub` model
- * @property {string} [employeeShortFIO] Short name of employee. Added by `org` model
- * @property {string} [employeeFullFIO] Full name of employee
+ * @property {string} [employeeShortFIO] Short name of employee. Added by `ub` model from uba_user.firstName. `org` model override it
+ * @property {string} [employeeFullFIO] Full name of employee. Added by `ub` model from uba_user.fullName. `org` model override it
  * @property {number} [employeeID] Employee ID
  * @property {string} [staffUnitFullName]
  * @property {string} [staffUnitName]
@@ -365,7 +365,7 @@ Session.reset = function (sessionID, userID) {
  */
 Session._getRBACInfo = function (userID) {
   let userInfo = Repository('uba_user')
-    .attrs(['name', 'uData', 'uPasswordHashHexa', 'lastPasswordChangeDate'])
+    .attrs(['name', 'uData', 'uPasswordHashHexa', 'lastPasswordChangeDate', 'firstName', 'lastName', 'fullName'])
     .selectById(userID)
   if (!userInfo) throw new Error(`User with ID=${userID} not found`)
   let uData = {}
@@ -380,6 +380,8 @@ Session._getRBACInfo = function (userID) {
   uData.userID = userID
   uData.roleIDs = [UBA_COMMON.ROLES.EVERYONE.ID]
   uData.login = userInfo.name
+  uData.employeeShortFIO = userInfo.firstName || userInfo.name
+  if (userInfo.fullName) uData.employeeFullFIO = userInfo.fullName
   let result = {
     uData: uData,
     uPasswordHashHexa: userInfo.uPasswordHashHexa,